pyproject.toml
[tool.poetry]
name = "tbump"
version = "6.1.1"
description = "Bump software releases"
authors = ["Dimitri Merejkowsky <d.merej@gmail.com>"]
readme = "README.rst"

[tool.poetry.dependencies]
# Note: keep this in sync with .github/workflows/tests.yml
python = "^3.6"

attrs = "^19.3.0"
docopt = "^0.6.2"
cli-ui = "^0.10.3"
schema = "^0.7.1"
tomlkit = "^0.5.8"

[tool.poetry.dev-dependencies]
black = { version = "19.10b0", python = '^3.6' }
pytest = "^6.0.1"
pytest-mock = "^2.0.0"
pytest-cov = "^2.8.1"
codecov = "^2.0.15"
mypy = "^0.790"
flake8 = "^3.7.9"
flake8-comprehensions = "^3.2.2"
safety = "^1.9.0"

[tool.poetry.scripts]
tbump = "tbump.main:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
